# docs/openapi.yml
basePath: /kbs/v1
definitions:
...
    User:
        properties:
            password:
                example: testPassword
                type: string
                x-go-name: Password
            permissions:
                description: Roles associated with an user
                example:
                    - users:create
                    - users:search
                items:
                    type: string
                type: array
                x-go-name: Permissions
            username:
                example: testUser
                type: string
                x-go-name: Username
        required:
            - username
            - password
            - permissions
        type: object
        x-go-package: intel/kbs/v1/model
    UserResponse:
        properties:
            created_at:
                description: Asset creation time
                example: "0001-01-01T00:00:00Z"
                format: date-time
                type: string
                x-go-name: CreatedAt
            id:
                description: Universal Unique IDentifier of the user
                example: 8170574b-a783-4357-747f-3e52922f2ed8
                format: uuid
                type: string
                x-go-name: ID
            permissions:
                description: Roles associated with an user
                example:
                    - users:create
                    - users:search
                items:
                    type: string
                type: array
                x-go-name: Permissions
            updated_at:
                description: Asset modification time
                example: "0001-01-01T00:00:00Z"
                format: date-time
                type: string
                x-go-name: UpdatedAt
            username:
                description: Specified Username for whom User ID is created
                example: testUser
                type: string
                x-go-name: Username
        type: object
        x-go-package: intel/kbs/v1/model
host: kbs.com:9443
info:
    description: |-
        The Key Broker Service enables key distribution using platform trust to authorize key transfers. By retaining image decryption keys,
        unauthorized access can be prevented, even from within the Cloud Service Provider. It interfaces with a backend key management system(KMIP)
        to create, delete and retrieve keys, while providing a user defined policy for key retrieval based on Intel® hardware root of trust.
    license:
        name: Copyright(C) 2022 Intel Corporation. All Rights Reserved.
    title: Key Broker Service
    version: 0.3.1
paths:
...
    /users:
        get:
            description: |
                Searches for users with a given set of filter criterias.
                Returns - The collection of serialized UserResponse Go struct objects.
            operationId: SearchUser
            parameters:
                - description: the name of the user want to search.
                  in: query
                  name: username
                  type: string
                - description: Accept header
                  enum:
                    - application/json
                  in: header
                  name: Accept
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Successfully retrieved the users.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Invalid values for request params
                "401":
                    description: Request Unauthorized
                "415":
                    description: Invalid Accept Header in Request
                "500":
                    description: Internal server error
            security:
                - bearerAuth: []
            tags:
                - User
            x-permissions: users:search
            x-sample-call-endpoint: https://kbs.com:9443/kbs/v1/users?username=keysAdmin
            x-sample-call-output: |
                {
                  "id": "9acad9da-4ef0-4865-9426-f9c5a8be4d62",
                  "created_at": "2022-09-16T03:13:09.503698068Z",
                  "updated_at": "0001-01-01T00:00:00Z",
                  "username": "keysAdmin",
                  "permissions": [
                  "users:create",
                  "keys:create"
                 ]
                }
        post:
            consumes:
                - application/json
            description: |
                Creates a user with the given username, password and API permissions

                The serialized User Go struct object represents the content of the request body.

                 | Attribute   | Description |
                 |-------------|-------------|
                 | username    | Name of the user |
                 | password    | The password of the user |\
                 | permissions | The KBS REST API permissions in ["{KBS_API}:{CRUD_permissions}"] format. Supported KBS API's are users, keys, key-transfer-policies. Supported CRUD_permissions are create, delete, search and update|
            operationId: CreateUser
            parameters:
                - in: body
                  name: request body
                  required: true
                  schema:
                    $ref: '#/definitions/User'
                - description: Content-Type header
                  enum:
                    - application/json
                  in: header
                  name: Content-Type
                  required: true
                  type: string
                - description: Accept header
                  enum:
                    - application/json
                  in: header
                  name: Accept
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Successfully created a user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Invalid request body provided
                "401":
                    description: Request Unauthorized
                "415":
                    description: Invalid Accept Header in Request
                "500":
                    description: Internal server error
            security:
                - bearerAuth: []
            tags:
                - User
            x-permissions: users:create
            x-sample-call-endpoint: https://kbs.com:9443/kbs/v1/users
            x-sample-call-input: |
                {
                   "username": "testUser",
                   "password": "testUserPassword",
                   "permissions": ["users:create", "keys:create"]
                }
            x-sample-call-output: |-
                {
                  "id": "9acad9da-4ef0-4865-9426-f9c5a8be4d62",
                  "created_at": "2022-09-16T03:13:09.503698068Z",
                  "updated_at": "0001-01-01T00:00:00Z",
                  "username": "keysAdmin",
                  "permissions": [
                  "users:create",
                  "keys:create"
                 ]
                }
    /users/{id}:
        delete:
            description: |
                Deletes a user account and associated roles.
            operationId: DeleteUser
            parameters:
                - description: Unique ID of the user.
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "204":
                    description: Successfully deleted the user.
                "401":
                    description: Request Unauthorized
                "404":
                    description: user record not found
                "500":
                    description: Internal server error
            security:
                - bearerAuth: []
            tags:
                - User
            x-permissions: users:delete
            x-sample-call-endpoint: https://kbs.com:9443/kbs/v1/users/9acad9da-4ef0-4865-9426-f9c5a8be4d62
        get:
            description: |
                Retrieves a user information for the provided user ID.
                Returns - The serialized UserResponse Go struct object that was retrieved.
            operationId: RetrieveUser
            parameters:
                - description: Unique ID of the user.
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Accept header
                  enum:
                    - application/json
                  in: header
                  name: Accept
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Successfully retrieved the user information.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "401":
                    description: Request Unauthorized
                "404":
                    description: User record not found
                "415":
                    description: Invalid Accept Header in Request
                "500":
                    description: Internal server error
            security:
                - bearerAuth: []
            tags:
                - User
            x-permissions: users:search
            x-sample-call-endpoint: https://kbs.com:9443/kbs/v1/users/9acad9da-4ef0-4865-9426-f9c5a8be4d62
            x-sample-call-output: |-
                {
                  "id": "9acad9da-4ef0-4865-9426-f9c5a8be4d62",
                  "created_at": "2022-09-16T03:13:09.503698068Z",
                  "updated_at": "0001-01-01T00:00:00Z",
                  "username": "keysAdmin",
                  "permissions": [
                  "users:create",
                  "keys:create"
                 ]
                }
        put:
            consumes:
                - application/json
            description: |
                Updates a user with username, password and API permissions provided in the body of the request for a user with ID in the path parameter

                The serialized User Go struct object represents the content of the request body.

                 | Attribute   | Description |
                 |-------------|-------------|
                 | username    | Name of the user |
                 | password    | The password of the user |\
                 | permissions | The KBS REST API permissions in ["{KBS_API}:{CRUD_permissions}"] format. Supported KBS API's are users, keys, key-transfer-policies. Supported CRUD_permissions are create, delete, search and update|
            operationId: UpdateUser
            parameters:
                - description: Unique ID of the user.
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: request body
                  required: true
                  schema:
                    $ref: '#/definitions/User'
                - description: Content-Type header
                  enum:
                    - application/json
                  in: header
                  name: Content-Type
                  required: true
                  type: string
                - description: Accept header
                  enum:
                    - application/json
                  in: header
                  name: Accept
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Successfully created a user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Invalid request body provided
                "401":
                    description: Request Unauthorized
                "415":
                    description: Invalid Accept Header in Request
                "500":
                    description: Internal server error
            security:
                - bearerAuth: []
            tags:
                - User
            x-permissions: users:update
            x-sample-call-endpoint: https://kbs.com:9443/kbs/v1/users/9acad9da-4ef0-4865-9426-f9c5a8be4d62
            x-sample-call-input: |
                {
                   "username": "updatedUsername",
                   "password": "testUserPassword",
                   "permissions": ["users:create", "keys:create"]
                }
            x-sample-call-output: |-
                {
                  "id": "9acad9da-4ef0-4865-9426-f9c5a8be4d62",
                  "created_at": "2022-09-16T03:13:09.503698068Z",
                  "updated_at": "0001-01-01T00:00:00Z",
                  "username": "updatedUsername",
                  "permissions": [
                  "users:create",
                  "keys:create"
                 ]
                }
...
securityDefinitions:
    bearerAuth:
        description: Enter your bearer token in the format <Bearer Token>
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"